+++ compute/src/server.rs
@@ -1,25 +1,23 @@
 use std::error::Error;
 use std::sync::Arc;
-use std::sync::mpsc::Sender;
 
 use grpcio;
 use grpcio::{RpcStatus, RpcStatusCode};
-use thread_local::ThreadLocal;
 
 use ekiden_compute_api::{CallContractRequest, CallContractResponse, Compute,
                          WaitContractCallRequest, WaitContractCallResponse};
 use ekiden_core::bytes::H256;
 use ekiden_core::futures::Future;
-use ekiden_core::futures::sync::oneshot;
 
+use super::consensus::ConsensusFrontend;
 use super::instrumentation;
-use super::worker::{Command, Worker};
+use super::worker::Worker;
 
 struct ComputeServiceInner {
     /// Worker.
     worker: Arc<Worker>,
-    /// Thread-local channel for submitting commands to the worker.
-    tl_command_sender: ThreadLocal<Sender<Command>>,
+    /// Consensus frontend.
+    consensus_frontend: Arc<ConsensusFrontend>,
     /// Instrumentation objects.
     ins: instrumentation::HandlerMetrics,
 }
@@ -31,21 +29,14 @@ pub struct ComputeService {
 
 impl ComputeService {
     /// Create new compute server instance.
-    pub fn new(worker: Arc<Worker>) -> Self {
+    pub fn new(worker: Arc<Worker>, consensus_frontend: Arc<ConsensusFrontend>) -> Self {
         ComputeService {
             inner: Arc::new(ComputeServiceInner {
                 worker,
-                tl_command_sender: ThreadLocal::new(),
+                consensus_frontend,
                 ins: instrumentation::HandlerMetrics::new(),
             }),
         }
-    }
-
-    /// Get thread-local command sender.
-    fn get_command_sender(&self) -> &Sender<Command> {
-        self.inner
-            .tl_command_sender
-            .get_or(|| Box::new(self.inner.worker.get_command_sender()))
     }
 }
 
@@ -60,14 +51,12 @@ impl Compute for ComputeService {
         self.inner.ins.reqs_received.inc();
         let _client_timer = self.inner.ins.req_time_client.start_timer();
 
-        // Send command to worker thread.
-        let (response_sender, response_receiver) = oneshot::channel();
-        self.get_command_sender()
-            .send(Command::RpcCall(
-                rpc_request.take_payload(),
-                response_sender,
-            ))
-            .unwrap();
+        // Send command to worker thread and request any generated batches to be handled
+        // by our consensus frontend.
+        let response_receiver = self.inner.worker.rpc_call(
+            rpc_request.take_payload(),
+            self.inner.consensus_frontend.clone(),
+        );
 
         // Prepare response future.
         let f = response_receiver.then(|result| match result {
@@ -105,26 +94,18 @@ impl Compute for ComputeService {
         }
 
         // Send command to worker thread.
-        let (response_sender, response_receiver) = oneshot::channel();
-        self.get_command_sender()
-            .send(Command::SubscribeCall(
-                H256::from(request.get_call_id()),
-                response_sender,
-            ))
-            .unwrap();
+        let response_receiver = self.inner
+            .consensus_frontend
+            .subscribe_call(H256::from(request.get_call_id()));
 
         // Prepare response future.
         let f = response_receiver.then(|result| match result {
-            Ok(Ok(response)) => {
+            Ok(response) => {
                 let mut rpc_response = WaitContractCallResponse::new();
                 rpc_response.set_output(response);
 
                 sink.success(rpc_response)
             }
-            Ok(Err(error)) => sink.fail(RpcStatus::new(
-                RpcStatusCode::Internal,
-                Some(error.description().to_owned()),
-            )),
             Err(error) => sink.fail(RpcStatus::new(
                 RpcStatusCode::Internal,
                 Some(error.description().to_owned()),