+++ consensus/dummy/src/lib.rs
@@ -1,4 +1,7 @@
 //! Ekiden dummy consensus backend.
+#[macro_use]
+extern crate log;
+
 extern crate ekiden_common;
 extern crate ekiden_consensus_base;